pyproject.toml
[project]
name = "study"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-genai>=1.57.0",
    "google-generativeai>=0.8.6",
    "openai>=2.15.0",
    "python-dotenv>=1.2.1",
    "watchdog>=6.0.0",
]
